refactor: centralizacao das entidades fakes
