+++ src/Application.Models/Customer/Requests/CreateCustomerRequest.cs
@@ -34,7 +34,7 @@ public class CreateCustomerRequest
         Number = number;
     }
 
-    public string FullName { get; }
+    public string FullName { get; set; }
     public string Email { get; set; }
     public string EmailConfirmation { get; set; }
     public string Cpf { get; set; }
