+++ src/UnitTests/EntitiesFake/CustomerBankInfos/CustomerBankInfoResponseModel.cs
@@ -1,0 +1,21 @@
+﻿using Application.Models.CustomerBackInfo.Response;
+using Bogus;
+
+namespace UnitTests.EntitiesFake.CustomerBankInfos;
+
+public static class CustomerBankInfoResponseModel
+{
+
+    public static CustomerBankInfoResult BankInfoFake()
+    {
+        var customerBankInfoResult = new Faker<CustomerBankInfoResult>("pt_BR")
+            .CustomInstantiator(f => new CustomerBankInfoResult(
+                id: 1,
+                account: f.Random.Guid().ToString(),
+                accountBalance: 0,
+                customerId: 1
+            )).Generate();
+
+        return customerBankInfoResult;
+    }
+}
