+++ src/UnitTests/EntitiesFake/Portfolios/CreateInvestmentModel.cs
@@ -1,0 +1,37 @@
+﻿using Application.Models.Enum;
+using Application.Models.Portfolio.Request;
+using Bogus;
+using System;
+
+namespace UnitTests.EntitiesFake.Portfolios;
+
+public static class CreateInvestmentModel
+{
+    public static InvestmentRequest InvestmentFake()
+    {
+        var investmentRequest = new Faker<InvestmentRequest>("pt_BR")
+            .CustomInstantiator(f => new InvestmentRequest(
+                ProductId: f.Random.Long(1, 1),
+                PortfolioId: f.Random.Long(1, 1),
+                Quotes: f.Random.Int(1, 10),
+                ConvertedAt: DateTime.Now.AddDays(-7),
+                Direction: OrderDirection.Buy
+            )).Generate();
+
+        return investmentRequest;
+    }
+
+    public static InvestmentRequest InvestmentInvalid()
+    {
+        var investmentRequest = new Faker<InvestmentRequest>("pt_BR")
+            .CustomInstantiator(f => new InvestmentRequest(
+                ProductId: f.Random.Long(1, 1),
+                PortfolioId: f.Random.Long(1, 1),
+                Quotes: f.Random.Int(1, 10),
+                ConvertedAt: f.Date.Future(),
+                Direction: OrderDirection.Buy
+            )).Generate();
+
+        return investmentRequest;
+    }
+}
